Getting ESP32 to work.
This is a clone (from v1.10) of the work done at https://github.com/scottrcarlson/RNode_Firmware/ to port the
RNode code to ESP32. The ESP32 has flash EEPROM, so we need to limit the number of EEPROM.commit() . That required
a change in the KISS protocol the device is originally using.

You need to use this code with a patched version of rnodeconf which I will check into a separate repo.

TODO: port the enhancements the original author did with v1.16 as well.